Add success/error messages for login/register and delay navigation for proper redirect

+++ src/main/java/com/my_project/service/ReservationService.java
@@ -33,7 +33,7 @@ public class ReservationService {
             throw new RuntimeException("Room already booked!");
         }
 
-        User user = userRepository.findById(userId)
+        User user = userRepository.findById(Math.toIntExact(Long.valueOf(userId)))
                 .orElseThrow(() -> new RuntimeException("User not found"));
 
         Room room = roomRepository.findById(roomId)
